Fix attempt for bug in blog page
Last commit caused a bug which makes clicking on the link to the blog page from the portfolio page cause an internal server error.
The only things that were changed were added styling to both pages, and reversed the order of the blog posts.
Since the styling worked fine on the portfolio page, it is likely that reversing the order caused it.
So changing the posts back to the old format (oldest first) and seeing if it fixes the bug.